refactor: update to the latest approach
