+++ src/screens/about-me/index.ts
@@ -1,0 +1,1 @@
+export * from './about-me'
