+++ src/screens/home/home.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react'
+
+import { Meta } from '~/components/meta'
+import { MainLayout } from '~/layouts'
+
+import { Content } from './components/content'
+
+export function HomeScreen() {
+  return (
+    <MainLayout>
+      <Meta />
+      <Content />
+    </MainLayout>
+  )
+}
