+++ src/screens/projects/components/content/index.ts
@@ -1,0 +1,1 @@
+export * from './content'
